@@ -11,48 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^13.0.3",
-    "@angular/animations": "~13.0.2",
-    "@angular/cli": "~13.0.3",
-    "@angular/common": "~13.0.2",
-    "@angular/compiler": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.2",
-    "@angular/core": "~13.0.2",
-    "@angular/forms": "~13.0.2",
-    "@angular/language-service": "~13.0.2",
-    "@angular/localize": "^13.0.2",
-    "@angular/platform-browser": "~13.0.2",
-    "@angular/platform-browser-dynamic": "~13.0.2",
-    "@angular/router": "~13.0.2",
-    "@angular/service-worker": "~13.0.2",
+    "@angular-devkit/build-angular": "^15.0.4",
+    "@angular-devkit/schematics": "^15.0.4",
+    "@angular/animations": "^15.0.4",
+    "@angular/cli": "^15.0.4",
+    "@angular/common": "^15.0.4",
+    "@angular/compiler": "^15.0.4",
+    "@angular/compiler-cli": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/forms": "^15.0.4",
+    "@angular/language-service": "^15.0.4",
+    "@angular/localize": "^15.0.4",
+    "@angular/platform-browser": "^15.0.4",
+    "@angular/platform-browser-dynamic": "^15.0.4",
+    "@angular/router": "^15.0.4",
+    "@angular/service-worker": "^15.0.4",
     "materialize-css": "^1.0.0-rc.2",
-    "rxjs": "~6.5.4",
+    "rxjs": "^6.5.4",
     "tslib": "^2.0.0",
-    "typescript": "~4.4.4",
-    "zone.js": "~0.11.4"
+    "typescript": "^4.8.4",
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "13.0.1",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/schematics": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-eslint/builder": "^15.1.0",
+    "@angular-eslint/eslint-plugin": "^15.1.0",
+    "@angular-eslint/eslint-plugin-template": "^15.1.0",
+    "@angular-eslint/schematics": "^15.1.0",
+    "@angular-eslint/template-parser": "^15.1.0",
+    "@types/jasmine": "^3.6.0",
+    "@types/jasminewd2": "^2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "codelyzer": "^6.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@typescript-eslint/parser": "^5.3.0",
+    "codelyzer": "^0.0.28",
     "eslint": "^8.2.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.4",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
+    "jasmine-core": "^3.6.0",
+    "jasmine-spec-reporter": "^5.0.0",
+    "karma": "^6.3.4",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^3.0.2",
+    "karma-jasmine": "^4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0"
+    "protractor": "^7.0.0",
+    "ts-node": "^8.3.0",
+    "tslint": "^6.1.0"
   }
 }
